move search components to subfolder
