+++ src/mdxGlobalComponents.js
@@ -107,7 +107,6 @@ import { Question } from './components/Question'
 import { Quote2 } from './components/Quote2'
 import { Roadmap } from './components/Roadmap'
 import { RocketHorizontalRule } from './components/RocketHorizontalRule'
-import { SearchBox } from './components/SearchBox'
 import { SectionLinks } from './components/SectionLinks'
 import { DesignedForYourStackBlock } from './components/Sections/DesignedForYourStackBlock'
 import { FeaturedSectionTextLeft } from './components/Sections/FeaturedSectionTextLeft'
@@ -240,7 +239,6 @@ export const shortcodes = {
     Quote2,
     Roadmap,
     RocketHorizontalRule,
-    SearchBox,
     SectionLinks,
     DesignedForYourStackBlock,
     FeaturedSectionTextLeft,
