+++ src/pages/using-posthog.tsx
@@ -7,7 +7,7 @@ import { docs } from '../sidebars/sidebars.json'
 import { graphql } from 'gatsby'
 import ProductIcons from 'components/ProductIcons'
 import { GatsbyImage, getImage } from 'gatsby-plugin-image'
-import SearchBox from 'components/SearchBox'
+import SearchBox from 'components/Search/SearchBox'
 import { Close } from 'components/Icons/Icons'
 import { CallToAction } from 'components/CallToAction'
 import Modal from 'components/Modal'
